refactor: ProjectForm and useProject(id)
