+++ pages/dashboard/projects/[id]/index.tsx
@@ -5,6 +5,7 @@ import Layout from "../../../../components/layout";
 import { Table, Tbody, Tr, Td, TableContainer } from "@chakra-ui/react";
 import { useEffect, useState } from "react";
 import { Project } from "../../../../db/models/project";
+import { useProject } from "../../../../lib/project-helpers";
 
 const ProjectTable = ({ project }: { project: Project }) => {
   return (
@@ -45,39 +46,10 @@ const ProjectTable = ({ project }: { project: Project }) => {
   );
 };
 
-export default function PostPage() {
+export default function ProjectPage() {
   const router = useRouter();
 
-  const [project, setProject] = useState<Project>();
-  const [isLoading, setLoading] = useState(false);
-  const [error, setError] = useState<string>();
-
-  useEffect(() => {
-    const getProject = async () => {
-      setLoading(true);
-
-      const id = parseInt(router.query.id as string) as number;
-      if (!id) return;
-
-      try {
-        const result = await fetch(`/api/v1/projects/${id}`);
-
-        const projectResult = await result.json();
-        if (projectResult.status === 200) {
-          setProject(projectResult);
-        } else {
-          setError(projectResult.error);
-        }
-      } catch (ex) {
-        console.error(ex);
-        setError("An error occurred when fetching the project information.");
-      }
-
-      setLoading(false);
-    };
-
-    getProject();
-  }, [router]);
+  const { project, error, isLoading } = useProject(router.query.id as string);
 
   return (
     <Layout>
